Env Changes for release 0.0.49
