+++ assets/training/finetune_acft_hf_nlp/environments/data_import/context/Dockerfile
@@ -1,0 +1,17 @@
+# openmpi image
+FROM mcr.microsoft.com/azureml/openmpi4.1.0-ubuntu20.04:{{latest-image-tag}}
+
+USER root
+
+# sudo is expected by Singularity inside the image
+RUN apt-get update && ACCEPT_EULA=Y apt-get -y upgrade && apt-get install sudo
+RUN pip install --upgrade pip
+
+COPY requirements.txt .
+
+RUN pip install -r requirements.txt --no-cache-dir
+
+# The below file is required for baking the code into the environment
+COPY data_import_run.py /azureml/data_import/run.py
+
+# dummy number to change when needing to force rebuild without changing the definition: 1
