+++ assets/training/finetune_acft_hf_nlp/environments/data_import/context/requirements.txt
@@ -1,0 +1,2 @@
+azureml_acft_common_components=={{latest-pypi-version}}
+azureml-acft-contrib-hf-nlp=={{latest-pypi-version}}
