+++ assets/training/finetune_acft_hf_nlp/environments/data_import/environment.yaml
@@ -1,0 +1,12 @@
+image:
+  name: azureml/curated/acft-hf-nlp-data-import
+  os: linux
+  context:
+    dir: context
+    dockerfile: Dockerfile
+    template_files:
+    - Dockerfile
+    - requirements.txt
+  publish:
+    location: mcr
+    visibility: public
